refactor: move LocalUpgradeInfo into a nodule
