+++ src/backend/lib/constants.js
@@ -1,5 +1,7 @@
 const DISCOVERY_KEY = "mapeo-upgrade";
+const UPGRADE_INFO_FILENAME = "upgrades.json";
 
 module.exports = {
-  DISCOVERY_KEY
+  DISCOVERY_KEY,
+  UPGRADE_INFO_FILENAME
 };
